Create a generic prefab SO (WIP). Fire an event when a setting is correctly configured in the editor and have other items respond to it (WIP).
