+++ Assets/Digital Painting/Data/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: feec48f059b19614287334d00d9f6f48
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
