+++ Assets/Digital Painting/ScriptableObjects/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7275da5d65d6ab459528fec34c440c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
