+++ Assets/Digital Painting/ScriptableObjects/Events/AgentWithFocusChangedGameEvent.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f03dc59164fb48e4d85e53599082025d, type: 3}
+  m_Name: AgentWithFocusChangedGameEvent
+  m_EditorClassIdentifier: 
+  DeveloperDescription:
+    _value: 
+  _debugValue: {fileID: 114301072202632092, guid: 4a45380c1234c7d418c8952744e3af01,
+    type: 3}
