+++ Assets/Digital Painting/ScriptableObjects/Events/AgentWithFocusChangedGameEvent.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a71f6489d3f89e3419d414d1e44fda34
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
