+++ Assets/Digital Painting/ScriptableObjects/Validation/AbstractSettingSO.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using UnityEditor;
+﻿using ScriptableObjectArchitecture;
+using System;
 using UnityEngine;
-using wizardscode.utility;
 
 namespace wizardscode.validation
 {
