+++ Assets/Digital Painting/ScriptableObjects/Validation/Base.meta
