+++ Assets/Digital Painting/ScriptableObjects/Validation/Base/FloatSettingSO.cs
