+++ Assets/Digital Painting/ScriptableObjects/Validation/Base/PrefabSettingSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e30612029f71a8e4d8cee30d3b90c5bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
