+++ Assets/Digital Painting/ScriptableObjects/Validation/Lighting/SunSettingSO.cs
@@ -6,17 +6,19 @@ using UnityEngine;
 namespace wizardscode.validation
 {
     [CreateAssetMenu(fileName = "SunSettingSO", menuName = "Wizards Code/Validation/Sun Prefab")]
-    public class SunSettingSO : GenericSettingSO<Light>
+    public class SunSettingSO : PrefabSettingSO<Light>
     {
-        protected override Light ActualValue
+        protected override UnityEngine.Object ActualValue
         {
             get { return RenderSettings.sun; }
-            set { RenderSettings.sun = value; }
+            set { RenderSettings.sun = ((GameObject)value).GetComponent<Light>() ; }
         }
 
+        /*
         public override void Fix()
         {
             RenderSettings.sun = GetFirstInstanceInScene().GetComponent<Light>();
         }
+        */
     }
 }
